move new files to a different folder, to avoid dependency cycle.
